tests/wpt/web-platform-tests/XMLHttpRequest/setrequestheader-case-insensitive.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setRequestHeader() - headers that differ in case</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-setrequestheader()-method" data-tested-assertations="/following::ol/li[6] /following::ol/li[7]" />  
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("POST", "resources/inspect-headers.py?filter_value=t1, t2, t3", false)
        client.setRequestHeader("x-test", "t1")
        client.setRequestHeader("X-TEST", "t2")
        client.setRequestHeader("X-teST", "t3")
        client.send(null)
        assert_equals(client.responseText, "x-test,")
      })
    </script>
  </body>
</html>
